backend/foodgram_backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
django-filter==24.2
Pillow==9.3.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
python-dotenv==1.0.1
djoser==2.2.2
gunicorn==20.1.0